+++ shared/src/androidMain/kotlin/com/alim/securevault/Platform.android.kt
@@ -1,9 +1,0 @@
-package com.alim.securevault
-
-import android.os.Build
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()