+++ shared/src/commonMain/kotlin/com/alim/securevault/GreetingUtil.kt
@@ -1,4 +1,0 @@
-package com.alim.securevault
-
-fun sayHello(to: String): String =
-    "Hello, $to!"