+++ shared/src/commonMain/kotlin/com/alim/securevault/Platform.kt
@@ -1,7 +1,0 @@
-package com.alim.securevault
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform